add BaseAsyncStream and refactor stream classes to derive from this
