+++ src/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/HttpContentReadStream.cs
@@ -11,6 +11,8 @@ namespace System.Net.Http
 {
     internal partial class HttpConnection
     {
+        // TODO: This becomes BaseReadOnlyStream (or rather, gets refactored)
+
         internal abstract class HttpContentReadStream : HttpContentStream
         {
             private int _disposed; // 0==no, 1==yes
@@ -22,32 +24,7 @@ namespace System.Net.Http
             public sealed override bool CanRead => true;
             public sealed override bool CanWrite => false;
 
-            public sealed override void Flush() { }
-            public sealed override Task FlushAsync(CancellationToken cancellationToken) =>
-                cancellationToken.IsCancellationRequested ?
-                    Task.FromCanceled(cancellationToken) :
-                    Task.CompletedTask;
-
-            public sealed override void WriteByte(byte value) => throw new NotSupportedException();
-            public sealed override void Write(byte[] buffer, int offset, int count) => throw new NotSupportedException();
-            public sealed override void Write(ReadOnlySpan<byte> source) => throw new NotSupportedException();
-            public sealed override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken) => throw new NotSupportedException();
             public sealed override ValueTask WriteAsync(ReadOnlyMemory<byte> destination, CancellationToken cancellationToken) => throw new NotSupportedException();
-
-            public sealed override int Read(byte[] buffer, int offset, int count)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-                return ReadAsync(new Memory<byte>(buffer, offset, count), CancellationToken.None).GetAwaiter().GetResult();
-            }
-
-            public sealed override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-                return ReadAsync(new Memory<byte>(buffer, offset, count), cancellationToken).AsTask();
-            }
-
-            public sealed override void CopyTo(Stream destination, int bufferSize) =>
-                CopyToAsync(destination, bufferSize, CancellationToken.None).GetAwaiter().GetResult();
 
             public virtual bool NeedsDrain => false;
