+++ src/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/HttpContentWriteStream.cs
@@ -18,27 +18,12 @@ namespace System.Net.Http
             public sealed override bool CanRead => false;
             public sealed override bool CanWrite => true;
 
-            public sealed override void Flush() => FlushAsync().GetAwaiter().GetResult();
-
-            public sealed override int Read(byte[] buffer, int offset, int count) => throw new NotSupportedException();
-
-            public sealed override void Write(byte[] buffer, int offset, int count) =>
-                WriteAsync(buffer, offset, count, CancellationToken.None).GetAwaiter().GetResult();
-
-            // The token here is ignored because it's coming from SendAsync and the only operations
-            // here are those that are already covered by the token having been registered with
-            // to close the connection.
-
-            public sealed override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken ignored)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-                return WriteAsync(new ReadOnlyMemory<byte>(buffer, offset, count), ignored).AsTask();
-            }
-
             public sealed override Task FlushAsync(CancellationToken ignored) =>
                 _connection.FlushAsync().AsTask();
 
             public abstract Task FinishAsync();
+
+            public sealed override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken) => throw new NotSupportedException();
         }
     }
 }
